add author and working on test-action-log

+++ test/test-object-manager.cc
@@ -1,6 +1,6 @@
 /* -*- Mode: C++; c-file-style: "gnu"; indent-tabs-mode:nil -*- */
 /*
- * Copyright (c) 2012 University of California, Los Angeles
+ * Copyright(c) 2015 University of California, Los Angeles
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 as
@@ -17,6 +17,7 @@
  *
  * Author: Alexander Afanasyev <alexander.afanasyev@ucla.edu>
  *	   Zhenkai Zhu <zhenkai@cs.ucla.edu>
+ *	   Lijing Wang <wanglj11@mails.tsinghua.edu.cn>
  */
 
 #include "logging.h"
@@ -30,46 +31,63 @@
 #include <boost/make_shared.hpp>
 #include <iostream>
 #include <iterator>
+#include <thread>
 
-INIT_LOGGER ("Test.ObjectManager");
+INIT_LOGGER("Test.ObjectManager");
 
-using namespace Ccnx;
+using namespace ndn;
 using namespace std;
 using namespace boost;
 namespace fs = boost::filesystem;
 
 BOOST_AUTO_TEST_SUITE(TestObjectManager)
 
-BOOST_AUTO_TEST_CASE (ObjectManagerTest)
+void listen(boost::shared_ptr<ndn::Face> face)
 {
-  INIT_LOGGERS ();
+  face->processEvents();
+  // do stuff...
+}
 
-  fs::path tmpdir = fs::unique_path (fs::temp_directory_path () / "%%%%-%%%%-%%%%-%%%%");
-  _LOG_DEBUG ("tmpdir: " << tmpdir);
-  Name deviceName ("/device");
+BOOST_AUTO_TEST_CASE(ObjectManagerTest)
+{
+  INIT_LOGGERS();
 
-  CcnxWrapperPtr ccnx = make_shared<CcnxWrapper> ();
-  ObjectManager manager (ccnx, tmpdir, "test-chronoshare");
+  fs::path tmpdir = fs::unique_path(fs::temp_directory_path() / "%%%%-%%%%-%%%%-%%%%");
+  _LOG_DEBUG("tmpdir: " << tmpdir);
+  Name deviceName("/device");
 
-  tuple<HashPtr,int> hash_semgents = manager.localFileToObjects (fs::path("test") / "test-object-manager.cc", deviceName);
+  
+  boost::shared_ptr<ndn::Face> face = boost::make_shared<ndn::Face>();
+  std::thread face_listeningThread(listen, face);
 
-  BOOST_CHECK_EQUAL (hash_semgents.get<1> (), 3);
+  ObjectManager manager(face, tmpdir, "test-chronoshare");
 
-  bool ok = manager.objectsToLocalFile (deviceName, *hash_semgents.get<0> (), tmpdir / "test.cc");
-  BOOST_CHECK_EQUAL (ok, true);
+  boost::tuple<ndn::ConstBufferPtr, int> hash_semgents = manager.localFileToObjects(fs::path("test") / "test-object-manager.cc", deviceName);
+
+  BOOST_CHECK_EQUAL(hash_semgents.get<1>(), 3);
+
+  bool ok = manager.objectsToLocalFile(deviceName, *hash_semgents.get<0>(), tmpdir / "test.cc");
+  BOOST_CHECK_EQUAL(ok, true);
 
   {
-    fs::ifstream origFile (fs::path("test") / "test-object-manager.cc");
-    fs::ifstream newFile (tmpdir / "test.cc");
+    fs::ifstream origFile(fs::path("test") / "test-object-manager.cc");
+    fs::ifstream newFile(tmpdir / "test.cc");
 
     istream_iterator<char> eof,
-      origFileI (origFile),
-      newFileI (newFile);
+      origFileI(origFile),
+      newFileI(newFile);
 
-    BOOST_CHECK_EQUAL_COLLECTIONS (origFileI, eof, newFileI, eof);
+    BOOST_CHECK_EQUAL_COLLECTIONS(origFileI, eof, newFileI, eof);
   }
 
-  remove_all (tmpdir);
+  if (exists(tmpdir))
+  {
+    std::cout << "Clear ALLLLLLLL" << std::endl;
+    face->shutdown();
+    face_listeningThread.detach();
+    remove_all(tmpdir);
+  }
+
 }
